test: 테스트를 도메인 기준으로 분리
